refactor: remove framework prefix and rename directories
