+++ internal/resources/metal/ssh_key/resource.go
