+++ internal/resources/metal/ssh_key/schema_common.go
